requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
yt-dlp>=2023.11.16
requests>=2.31.0
validators>=0.20.0
pillow>=9.0.0